+++ CHANGELOG.md
@@ -35,6 +35,9 @@
 - [BREAKING] Redid the implementation of how "Applied Effect" Power Roll Effects work. Now, instead of several sets, you can add an entire effect and it will have multiple properties to configure per tier, such as the Potency Condition, if/how it overrides the end, and any additional properties like "stackable".
 - [BREAKING] Moved recoveries from `system.hero.recoveries` to `system.recoveries`.
 - [BREAKING] Character max recoveries and stamina are now derived from their class item (#627).
+- Certain item types (Class, Subclass, Ancestry, Career, Kit) can no longer be created directly in world. Instead, they must be created inside a compendium. (#716)
+  - This enforces good practices for data creation.
+  - The buttons on the character sheet header now open the Ancestry, Background, and Class compendiums.
 
 ### Fixed